+++ sdk/keyvault/keyvault-certificates/package.json
@@ -73,11 +73,11 @@
   "//metadata": {
     "constantPaths": [
       {
-        "path": "src/core/keyVaultClientContext.ts",
+        "path": "src/generated/keyVaultClientContext.ts",
         "prefix": "packageVersion"
       },
       {
-        "path": "src/core/utils/constants.ts",
+        "path": "src/generated/utils/constants.ts",
         "prefix": "SDK_VERSION"
       }
     ]
