+++ sdk/keyvault/keyvault-keys/src/cryptographyClient.ts
@@ -23,9 +23,9 @@ import {
 import { getTracer } from "@azure/core-tracing";
 import { Span } from "@opentelemetry/api";
 import { logger } from "./log";
-import { parseKeyvaultIdentifier } from "./core/utils";
-import { SDK_VERSION } from "./core/utils/constants";
-import { KeyVaultClient } from "./core/keyVaultClient";
+import { parseKeyvaultIdentifier } from "./generated/utils";
+import { SDK_VERSION } from "./generated/utils/constants";
+import { KeyVaultClient } from "./generated/keyVaultClient";
 import { challengeBasedAuthenticationPolicy } from "../../keyvault-common/src";
 import { createHash as cryptoCreateHash, createVerify, publicEncrypt } from "crypto";
 import * as constants from "constants";
